requirements.txt
confluent-kafka==2.8.2
psycopg[binary]==3.2.6
python-dotenv==1.0.1
pydantic==2.10.6
prometheus-client==0.21.1
pytest==8.3.5
